fixed get_unipreot_id_mapping_pairs caching logic

+++ uniprot.py
@@ -48,17 +48,20 @@ def get_uniprot_id_mapping_pairs(
   from_type and to_type can be obtained from:
     http://www.uniprot.org/faq/28#mapping-faq-table
   """
-  r = requests.post(
-      'http://www.uniprot.org/mapping/', 
-      params={
-        'from': from_type,
-        'to': to_type,
-        'format': 'tab',
-        'query': ' '.join(seqids)})
-  text = r.text
   if cache_fname:
-    with open(cache_fname, 'w') as f:
-      f.write(text)
+    if not os.path.isfile(cache_fname):
+      r = requests.post(
+          'http://www.uniprot.org/mapping/', 
+          params={
+            'from': from_type,
+            'to': to_type,
+            'format': 'tab',
+            'query': ' '.join(seqids)})
+      text = r.text
+      with open(cache_fname, 'w') as f:
+        f.write(text)
+    else:
+      text = open(cache_fname).read()
   if is_html(text):
     # failed call results in a HTML error reporting page
     print "Error in fetching metadata"
@@ -417,6 +420,7 @@ def get_metadata_with_some_seqid_conversions(seqids, cache_fname=None):
 
   uniprot_seqids = [entry['uniprot_id'] for entry in entries 
                     if 'uniprot_id' in entry]
+  print 'cache', cache_fname
   uniprot_dict = batch_uniprot_metadata(uniprot_seqids, cache_fname)
 
   result = {}
